http://bugs.activestate.com/show_bug.cgi?id=74097 file properties do not work properly
The prefs system requires that you do not make pref changes until the dialog is closing.  The file properties panel was making changes during the ok button command handler, which were then lost due to a subsequent update to the file prefs.  See koPrefWindow.js for a description of the pref panel api.

This patch fixes issues in changing prefs for a file.  There are still problems with using prefs on new files in both bugs.

r=trentm


git-svn-id: http://svn.openkomodo.com/repos/openkomodo/trunk@528 b87d16e7-d29f-4ee1-b5ec-817d645d335f