find: Fix "Zero Information Alerts" http://bugs.activestate.com/show_bug.cgi?id=65220
Update the shared find_function to (a) by default send UI messages about
find/replace status to the statusbar and (b) provide a callback for callers to
handle those messages. The coming new find dialog will use the latter to put
messages in a notification box on the find dialog itself. The statusbar default
is still important for things in cmd_findNext (F3), etc.

r=toddw


git-svn-id: http://svn.openkomodo.com/repos/openkomodo/trunk@600 b87d16e7-d29f-4ee1-b5ec-817d645d335f