find: Fix 'First "replace" if selection was used to set the find term is skipped' http://bugs.activestate.com/show_bug.cgi?id=27208
The "Replace" button in the find dialog does:
1. replace the last find hit
2. jump to the next find hit

The problem was that (1) was only considering the "last find hit" logged in the
find session (a find session is tracked for other reasons). I.e. the current
selection wasn't considered a "last find hit" because the find session hadn't
started yet.

Solution: When doing "Replace" and the find session hasn't started yet, attempt
to prime the find session with a hit considering only the current selection.

r=toddw


git-svn-id: http://svn.openkomodo.com/repos/openkomodo/trunk@698 b87d16e7-d29f-4ee1-b5ec-817d645d335f