Fix "scintilla's in separate windows don't appear on Mac OS X 10.4" http://bugs.activestate.com/show_bug.cgi?id=79727
This wasn't a problem on 10.5. *Something* changed with moz1.9/macosx10.4 such
that the window (i.e. `fPlatform->container`) isn't attached to the open
QuickDraw port for a while. At least that is Shane's guess.

Here we reset the SciMoz (`PlatformResetWindow`) if the container is null.
After a number of `PlatformSetWindow` calls from mozilla, eventually things
hookup.

r=shanec

(integrated from komodo (trunk) change 22670 by trentm)

git-svn-id: http://svn.openkomodo.com/repos/openkomodo/trunk@2103 b87d16e7-d29f-4ee1-b5ec-817d645d335f