fix: slow typing response (mac os x) http://bugs.activestate.com/show_bug.cgi?id=80683
This is to use asynchronous API calls to open files within Komodo. This works
around a plugin problem, where if a Komodo editor file was opened from an
event (such as a double-click, or keyboard event) then the SciMoz plugin was
not getting correctly destroyed when the file was closed, resulting in higher
CPU usage everytime a file was closed in Komodo.

This code breaks some of the Komodo API, such as "ko.open.URI", which is now
made to work asynchronously, and now does not return the new view which was
recently opened. If the view is required, the programmer should use the
ko.views.manager.doFileOpenAsync() method and pass a callback function argument.

All of the existing ko.views.manager API functions that create new views are
now deprecated, and a matching Async function has been created in it's place.

r=trentm
rn=


git-svn-id: http://svn.openkomodo.com/repos/openkomodo/trunk@2751 b87d16e7-d29f-4ee1-b5ec-817d645d335f