Add drag-and-drop scheme URLs into Komodo. Finishes http://bugs.activestate.com/show_bug.cgi?id=82632 r=toddw rn=Scheme files are now drag-and-droppable into Komodo. Three ways are supported:
1. Drag a URL from a web browser into Komodo.  If it ends with
a ".ksf" extension, Komodo will offer to either load it as
a scheme, or as a plain text file.

2. Specify a .ksf file on the command-line.

3. Simply open a .ksf file inside Komodo, using the usual
file-open commands.

Note that in case 1 Komodo will check the contents of schemes
from other machines before loading them.  If it suspects the
contents is not actually a Komodo scheme, it will load the
file into a Komodo editor view.  If you want to load it as
a scheme, you can now simply reopen the local edited view.

(integrated from komodo (trunk) change 27784 by ericp)

git-svn-id: http://svn.openkomodo.com/repos/openkomodo/trunk@3380 b87d16e7-d29f-4ee1-b5ec-817d645d335f