workaround: UI problems with the new notificationbox in komodo.xul http://bugs.activestate.com/show_bug.cgi?id=82863
To workaround the scrolling (on Mac with trackpad), flex/overflow, drop marker
for tab dragging, etc. problems caused by other placements and wrapping of the
notificationbox, it has been moved to be (a) on top of the editor box *and* the
sidebar boxes (so that the notificationbox appears on top of, e.g., the
toolbox) and (b) does NOT contain that element (necessary to get tab drop
markers to show (don't know why).

r=ericp/trentm

(integrated from komodo (trunk) change 28178 by trentm)

git-svn-id: http://svn.openkomodo.com/repos/openkomodo/trunk@3560 b87d16e7-d29f-4ee1-b5ec-817d645d335f