Tools: fix http://bugs.activestate.com/show_bug.cgi?id=87763 "Remove id fields from individual tools"
We now use the absolute path + timestamp to uniquely identify 
each tool.  If a path in the DB isn't found on a load, it's
removed from the database.  And of course, if a tool has
been edited on disk, it's reinitialized.

-This line, and those below, will be ignored--

M    koToolbox2.py

(integrated from komodo (trunk) change 55556 by ericp)

git-svn-id: http://svn.openkomodo.com/repos/openkomodo/trunk@5987 b87d16e7-d29f-4ee1-b5ec-817d645d335f