patchtree: support queues with ordering http://bugs.activestate.com/show_bug.cgi?id=101191
This lets patchsets to specify ordering of patches; just supply a file
named "series" in the patchset directory.  The contents should be the names
of the patches to apply, one on each line.

Patches may overlap; they will be applied in order (and unapplied in
reverse order).  It is no longer necessary to avoid changing the context of
other patches.

Patchsets without explicit ordering are applied in an undefined order.

(integrated from komodo (trunk) change 82806 by marky)

git-svn-id: http://svn.openkomodo.com/repos/openkomodo/trunk@13903 b87d16e7-d29f-4ee1-b5ec-817d645d335f