views: go to def: check for codeintel completion availability (r=toddw)
Make sure that go-to-def is actually acting on views that have ci cpln
available; otherwise, we will throw when attempting to complete the
codeintel callback and break the codeintel manager.

As a side effect, this constructs the completion UI handler if it should
exist but hasn't been created yet.

(integrated from komodo (trunk) change 83288 by marky)

git-svn-id: http://svn.openkomodo.com/repos/openkomodo/trunk@14062 b87d16e7-d29f-4ee1-b5ec-817d645d335f