Fixed an error in the Run service
When executing a command in the Run service of Komodo using a variable
like %F raises an error if the filepath has spaces in it.

Like an example, when working with the file which path is /home/user/my
project/my file.py and executing the python %F command through the run
service, an error is raised because the command tries to open that
exactly ath and not the /home/user/my\ project/my\ file.py.

Through some tests i can tell it works althought it hasen't been tested
on Komodo Edit because of some problems with the building process.