UI: Reduce top-padding of toolbar when maximized, fixing menu not being accessible at times
The 10px padding on the menu is slightly too much.  When you move the cursor to the top of the screen, the menu is 2px too low to click on, so you have to move back down slightly.  This change lowers the padding so that you can throw your mouse cursor to the top of the screen and click a menu immediately.
rn=