codeintel: Remove typo from codeintel.p.css (#3351)
I suppose this might be "included" as part of a bigger template but from running Komodo and looking at the startup logs it seems like this is also dynamically interpreted as a syntax error. I assume its not actually doing anything due to this so I think it should be removed.